removing pointless abstract class
